IG orchestration works Most tests work

+++ services/src/main/java/gov/nist/toolkit/services/server/orchestration/IgOrchestrationBuilder.groovy
@@ -1,27 +1,22 @@
 package gov.nist.toolkit.services.server.orchestration
 
 import gov.nist.toolkit.actorfactory.SimCache
-import gov.nist.toolkit.configDatatypes.SimulatorProperties
 import gov.nist.toolkit.actorfactory.client.SimId
 import gov.nist.toolkit.actorfactory.client.SimulatorConfig
 import gov.nist.toolkit.actortransaction.client.ActorType
-import gov.nist.toolkit.configDatatypes.client.PidBuilder
-import gov.nist.toolkit.configDatatypes.client.TransactionType
-import gov.nist.toolkit.configDatatypes.client.PatientError
-import gov.nist.toolkit.configDatatypes.client.PatientErrorList
-import gov.nist.toolkit.configDatatypes.client.PatientErrorMap
-import gov.nist.toolkit.configDatatypes.client.Pid
+import gov.nist.toolkit.configDatatypes.SimulatorProperties
+import gov.nist.toolkit.configDatatypes.client.*
 import gov.nist.toolkit.installation.Installation
-import gov.nist.toolkit.results.shared.SiteBuilder
 import gov.nist.toolkit.results.client.TestInstance
-import gov.nist.toolkit.services.client.IgOrchestrationRequest
-import gov.nist.toolkit.services.client.IgOrchestrationResponse
-import gov.nist.toolkit.services.client.RawResponse
+import gov.nist.toolkit.results.shared.SiteBuilder
+import gov.nist.toolkit.services.client.*
 import gov.nist.toolkit.services.server.RawResponseBuilder
 import gov.nist.toolkit.services.server.ToolkitApi
 import gov.nist.toolkit.session.server.Session
 import gov.nist.toolkit.simcommon.client.config.SimulatorConfigElement
+import gov.nist.toolkit.sitemanagement.client.Site
 import groovy.transform.TypeChecked
+
 /**
  * Build environment for testing Initiating Gateway SUT.
  *
@@ -34,6 +29,7 @@ class IgOrchestrationBuilder {
     Util util
     List<SimulatorConfig> rgConfigs = []
     SimulatorConfig igConfig = null
+    OrchestrationProperties orchProps
 
     public IgOrchestrationBuilder(ToolkitApi api, Session session, IgOrchestrationRequest request) {
         this.api = api
@@ -46,13 +42,13 @@ class IgOrchestrationBuilder {
         try {
             IgOrchestrationResponse response = new IgOrchestrationResponse()
             Map<String, TestInstanceManager> pidNameMap = [
-                    oneDocPid:  new TestInstanceManager(request, response, '15824'),
-                    twoDocPid:  new TestInstanceManager(request, response, '15825'),
-                    twoRgPid:  new TestInstanceManager(request, response, '15826'),
-                    registryErrorPid:  new TestInstanceManager(request, response, '15827'),
+                    oneDocPid       : new TestInstanceManager(request, response, '15824'),
+                    twoDocPid       : new TestInstanceManager(request, response, '15825'),
+                    twoRgPid        : new TestInstanceManager(request, response, '15826'),
+                    registryErrorPid: new TestInstanceManager(request, response, '15827'),
             ]
 
-            OrchestrationProperties orchProps = new OrchestrationProperties(session, request.userName, ActorType.INITIATING_GATEWAY, pidNameMap.keySet())
+            orchProps = new OrchestrationProperties(session, request.userName, ActorType.INITIATING_GATEWAY, pidNameMap.keySet())
 
             Pid oneDocPid = PidBuilder.createPid(orchProps.getProperty("oneDocPid"))
             Pid twoDocPid = PidBuilder.createPid(orchProps.getProperty("twoDocPid"))
@@ -64,35 +60,77 @@ class IgOrchestrationBuilder {
             response.setTwoRgPid(twoRgPid)
             response.setUnknownPid(registryErrorPid)
 
-            buildRGs(registryErrorPid)
+            List<String> simIds = buildRGs(registryErrorPid)
+
+            Site rg1Site = SimCache.getSite(session.getId(), simIds[0])
+            Site rg2Site = SimCache.getSite(session.getId(), simIds[1])
 
             String home0 = rgConfigs.get(0).get(SimulatorProperties.homeCommunityId).asString()
             String home1 = rgConfigs.get(1).get(SimulatorProperties.homeCommunityId).asString()
 
-            // Submit test data
-            util.submit(request.userName, SiteBuilder.siteSpecFromSimId(rgConfigs.get(0).id), new TestInstance("15807"), 'onedoc1', oneDocPid, home0)
-            util.submit(request.userName, SiteBuilder.siteSpecFromSimId(rgConfigs.get(0).id), new TestInstance("15807"), 'twodoc', twoDocPid, home0)
+            if (orchProps.updated()) {
+                // send necessary Patient ID Feed messages
+                request.setPifType(PifType.V2)
+                new PifSender(api, request.getUserName(), rg1Site.siteSpec(), orchProps).send(PifType.V2, pidNameMap)
+                new PifSender(api, request.getUserName(), rg2Site.siteSpec(), orchProps).send(PifType.V2, pidNameMap)
+
+                TestInstance testInstance15807 = TestInstanceManager.initializeTestInstance(request, new TestInstance('15807'))
+                MessageItem itemOneDoc1 = response.addMessage(testInstance15807, true, "")
+                MessageItem itemTwoDoc = response.addMessage(testInstance15807, true, "")
+                MessageItem itemOneDoc2 = response.addMessage(testInstance15807, true, "")
+                MessageItem itemOneDoc3 = response.addMessage(testInstance15807, true, "")
+                MessageItem itemRegistryError = response.addMessage(testInstance15807, true, "")
 
 
-            Map<String, String> params
-            params = [
-                    '$patientid$': twoRgPid.asString(),
-                    '$testdata_home$': home0,
-                    '$testdata_repid$': rgConfigs[0].getConfigEle(SimulatorProperties.repositoryUniqueId).asString()]
-            util.submit(request.userName, SiteBuilder.siteSpecFromSimId(rgConfigs.get(0).id), new TestInstance("15807"), 'onedoc2', params)
+                // Submit test data
+                try {
+                    util.submit(request.userName, SiteBuilder.siteSpecFromSimId(rgConfigs.get(0).id), new TestInstance("15807"), 'onedoc1', oneDocPid, home0)
+                } catch (Exception e) {
+                    itemOneDoc1.setMessage("Initialization of " + rgConfigs.get(0).id + " (section onedoc1) failed:\n" + e.getMessage());
+                    itemOneDoc1.setSuccess(false)
+                }
+                try {
+                    util.submit(request.userName, SiteBuilder.siteSpecFromSimId(rgConfigs.get(0).id), new TestInstance("15807"), 'twodoc', twoDocPid, home0)
+                } catch (Exception e) {
+                    itemTwoDoc.setMessage("Initialization of " + rgConfigs.get(0).id + " (section twodoc) failed:\n" + e.getMessage());
+                    itemTwoDoc.setSuccess(false)
+                }
 
-            params = [
-                    '$patientid$': twoRgPid.asString(),
-                    '$testdata_home$': home1,
-                    '$testdata_repid$': rgConfigs[1].getConfigEle(SimulatorProperties.repositoryUniqueId).asString()]
-            util.submit(request.userName, SiteBuilder.siteSpecFromSimId(rgConfigs.get(1).id), new TestInstance("15807"), 'onedoc3', params)
 
-            params = [
-                    '$patientid$': registryErrorPid.asString(),
-                    '$testdata_home$': home1,
-                    '$testdata_repid$': rgConfigs[1].getConfigEle(SimulatorProperties.repositoryUniqueId).asString()]
-            util.submit(request.userName, SiteBuilder.siteSpecFromSimId(rgConfigs.get(1).id), new TestInstance("15807"), 'registryErrorPid', params)
+                Map<String, String> params
+                params = [
+                        '$patientid$'     : twoRgPid.asString(),
+                        '$testdata_home$' : home0,
+                        '$testdata_repid$': rgConfigs[0].getConfigEle(SimulatorProperties.repositoryUniqueId).asString()]
+                try {
+                    util.submit(request.userName, SiteBuilder.siteSpecFromSimId(rgConfigs.get(0).id), new TestInstance("15807"), 'onedoc2', params)
+                } catch (Exception e) {
+                    itemOneDoc2.setMessage("Initialization of " + rgConfigs.get(0).id + " (section onedoc2) failed:\n" + e.getMessage());
+                    itemOneDoc2.setSuccess(false)
+                }
 
+                params = [
+                        '$patientid$'     : twoRgPid.asString(),
+                        '$testdata_home$' : home1,
+                        '$testdata_repid$': rgConfigs[1].getConfigEle(SimulatorProperties.repositoryUniqueId).asString()]
+                try {
+                    util.submit(request.userName, SiteBuilder.siteSpecFromSimId(rgConfigs.get(1).id), new TestInstance("15807"), 'onedoc3', params)
+                } catch (Exception e) {
+                    itemOneDoc3.setMessage("Initialization of " + rgConfigs.get(1).id + " (section onedoc3) failed:\n" + e.getMessage());
+                    itemOneDoc3.setSuccess(false)
+                }
+
+                params = [
+                        '$patientid$'     : registryErrorPid.asString(),
+                        '$testdata_home$' : home1,
+                        '$testdata_repid$': rgConfigs[1].getConfigEle(SimulatorProperties.repositoryUniqueId).asString()]
+                try {
+                    util.submit(request.userName, SiteBuilder.siteSpecFromSimId(rgConfigs.get(1).id), new TestInstance("15807"), 'registryError', params)
+                } catch (Exception e) {
+                    itemRegistryError.setMessage("Initialization of " + rgConfigs.get(1).id + " (section registryError) failed:\n" + e.getMessage());
+                    itemRegistryError.setSuccess(false)
+                }
+            }
 
             response.oneDocPid = oneDocPid
             response.twoDocPid = twoDocPid
@@ -100,6 +138,8 @@ class IgOrchestrationBuilder {
             response.unknownPid = registryErrorPid
             response.simulatorConfigs = rgConfigs
             response.igSimulatorConfig = igConfig
+            response.supportRG1 = SimCache.getSite(session.getId(), simIds[0])
+            response.supportRG2 = SimCache.getSite(session.getId(), simIds[1])
 
             return response
         } catch (Exception e) {
@@ -107,18 +147,31 @@ class IgOrchestrationBuilder {
         }
     }
 
-    void buildRGs(Pid unknownPid) {
+    List<String> buildRGs(Pid unknownPid) {
         // build and initialize remote communities
-        String id1 = 'rg1'
-        String id2 = 'rg2'
+        String id1 = 'community1'
+        String id2 = 'community2'
         SimId rgSimId1 = new SimId(request.userName, id1, ActorType.RESPONDING_GATEWAY.name, request.environmentName)
         SimId rgSimId2 = new SimId(request.userName, id2, ActorType.RESPONDING_GATEWAY.name, request.environmentName)
-        println "creating rg1 sim"
-        api.deleteSimulator(rgSimId1);
-        SimulatorConfig rgSimConfig1 = api.createSimulator(rgSimId1).getConfig(0)
-        println "creating rg2 sim"
-        api.deleteSimulator(rgSimId2);
-        SimulatorConfig rgSimConfig2 = api.createSimulator(rgSimId2).getConfig(0)
+        SimulatorConfig rgSimConfig1
+        SimulatorConfig rgSimConfig2
+        boolean reuse = false  // updated as we progress
+
+        if (!request.isUseExistingState()) {
+            api.deleteSimulator(rgSimId1);
+            api.deleteSimulator(rgSimId2);
+            orchProps.clear()
+        }
+
+        if (api.simulatorExists(rgSimId1)) {
+            rgSimConfig1 = api.getConfig(rgSimId1)
+            rgSimConfig2 = api.getConfig(rgSimId2)
+            reuse = true
+        } else {
+            rgSimConfig1 = api.createSimulator(rgSimId1).getConfig(0)
+            rgSimConfig2 = api.createSimulator(rgSimId2).getConfig(0)
+        }
+
 
         // this expects full server version of simulator config
         // this call makes the configuration available as a site for the test client
@@ -127,40 +180,46 @@ class IgOrchestrationBuilder {
 
         SimulatorConfigElement rgEle
 
-        // disable checking of Patient Identity Feed
-        rgEle = rgSimConfig1.getConfigEle(SimulatorProperties.VALIDATE_AGAINST_PATIENT_IDENTITY_FEED)
-        rgEle.setValue(false)
+        //
+        // Initialize both supporting Responding Gateways
+        //
 
-        // set fixed homeCommunityId
-        rgEle = rgSimConfig1.getConfigEle(SimulatorProperties.homeCommunityId)
-        rgEle.setValue('urn:oid:1.2.34.567.8.1')
+        if (!reuse) {
+            // disable checking of Patient Identity Feed
+//            rgEle = rgSimConfig1.getConfigEle(SimulatorProperties.VALIDATE_AGAINST_PATIENT_IDENTITY_FEED)
+//            rgEle.setValue(false)
 
-        // config rg1 to return XDSRegistryError for registryErrorPid query requests
-        rgEle = rgSimConfig1.getConfigEle(SimulatorProperties.errorForPatient)
-        PatientErrorMap pem = new PatientErrorMap()
-        PatientErrorList pel = new PatientErrorList()
-        PatientError pe = new PatientError()
-        pe.setPatientId(unknownPid)
-        pe.setErrorCode('XDSRegistryError')
-        pel.add(pe)
-        pem.put(TransactionType.XC_QUERY.name, pel)
-        rgEle.setValue(pem)
+            // set fixed homeCommunityId
+            rgEle = rgSimConfig1.getConfigEle(SimulatorProperties.homeCommunityId)
+            rgEle.setValue('urn:oid:1.2.34.567.8.1')
 
-        api.saveSimulator(rgSimConfig1)
+            // config rg1 to return XDSRegistryError for registryErrorPid query requests
+            rgEle = rgSimConfig1.getConfigEle(SimulatorProperties.errorForPatient)
+            PatientErrorMap pem = new PatientErrorMap()
+            PatientErrorList pel = new PatientErrorList()
+            PatientError pe = new PatientError()
+            pe.setPatientId(unknownPid)
+            pe.setErrorCode('XDSRegistryError')
+            pel.add(pe)
+            pem.put(TransactionType.XC_QUERY.name, pel)
+            rgEle.setValue(pem)
 
-        // disable checking of Patient Identity Feed
-        rgEle = rgSimConfig2.getConfigEle(SimulatorProperties.VALIDATE_AGAINST_PATIENT_IDENTITY_FEED)
-        rgEle.setValue(false)
-        // set fixed homeCommunityId
-        rgEle = rgSimConfig2.getConfigEle(SimulatorProperties.homeCommunityId)
-        rgEle.setValue('urn:oid:1.2.34.567.8.2')
-        api.saveSimulator(rgSimConfig2)
+            api.saveSimulator(rgSimConfig1)
+
+            // disable checking of Patient Identity Feed
+//            rgEle = rgSimConfig2.getConfigEle(SimulatorProperties.VALIDATE_AGAINST_PATIENT_IDENTITY_FEED)
+//            rgEle.setValue(false)
+            // set fixed homeCommunityId
+            rgEle = rgSimConfig2.getConfigEle(SimulatorProperties.homeCommunityId)
+            rgEle.setValue('urn:oid:1.2.34.567.8.2')
+            api.saveSimulator(rgSimConfig2)
+        }
 
         rgConfigs << rgSimConfig1
         rgConfigs << rgSimConfig2
 
         if (request.includeLinkedIG) {
-            // create initiating gateway
+            // create initiating gateway (SUT) to allow self test
             String igId = 'ig'
             SimId igSimId = new SimId(request.userName, igId, ActorType.INITIATING_GATEWAY.name, request.environmentName)
             igConfig = api.createSimulator(igSimId).getConfig(0);
@@ -171,5 +230,7 @@ class IgOrchestrationBuilder {
             rgs.setValue(rgConfigIds)
             api.saveSimulator(igConfig)
         }
+
+        return [rgSimId1.toString(), rgSimId2.toString()]
     }
 }
